requirements.txt
Mastodon.py
mecab-python3
